backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
aiosqlite==0.20.0
torch==2.11.0
torchvision==0.26.0
transformers==4.48.0
python-multipart==0.0.20
Pillow==12.2.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
ultralytics==8.3.59
